Generalise over the Swagger datatype.
This change modifies the previously introduced generalisation of the
Haskell datatype that holds the Swagger specification to an aeson Value.

By using an aeson Value, all ordering information of object fields is
being lost. This information is in principle preserved by aeson, when
going via Encoding rather than Value.

With this patch, we do not perform any aeson-related translation in the
servant-swagger-ui code, and simply use the original datatype. It is
servant's own responsibility to handle the actual conversion of the
datatype to JSON.

Based on https://github.com/haskell-servant/servant-swagger-ui/pull/89